Move edit-profile component to profile components
