+++ src/js/components/profile/edit-profile.js
@@ -1,10 +1,10 @@
-import { updateUser } from "../helper/api/putRequests/update-user.js";
-import { getFormData } from "../helper/get-form-data.js";
-import htmlUtilities from "../helper/html-utilities/index.js";
-import { DialogAlert } from "./alerts/dialog-alert.js";
-import { InputGroup } from "./input-group.js";
-import { AppLoader } from "./app-loader.js";
-import Modal from "../helper/modal/index.js";
+import { updateUser } from "../../helper/api/putRequests/update-user.js";
+import { getFormData } from "../../helper/get-form-data.js";
+import htmlUtilities from "../../helper/html-utilities/index.js";
+import { DialogAlert } from "../alerts/dialog-alert.js";
+import { InputGroup } from "../input-group.js";
+import { AppLoader } from "../app-loader.js";
+import Modal from "../../helper/modal/index.js";
 
 /**
  * Represents an `EditProfile` class, which is a modal component for editing user avatar and banner links.
